Codex/v54 unified composition (#19)
* feat: add V5.4 composition model and supervised curriculum

Add versioned composition C/G codecs, V54Model and all five named sizes.
Small defaults to three Unit layers; Nano remains an explicit competing
configuration. Preserve V5.3 codec IDs, defaults and seeded realizations.

Add curriculum-v54 with supervised-row as the current default and random-cell
as the first alternative. Version plan, checkpoint and terminal identities;
retain strict resume and scorer-only truth boundaries. Include the exact
owner-maintained V5.4 TeX snapshot and local examples.

Validation: 638 restoration/model/curriculum tests and 30 CLI/catalog tests
passed; scoped Ruff and staged diff checks passed. Small/Nano CPU optimizer
steps and the installed CLI V5.4 preflight passed, including exact next-update
checkpoint parity. Independent codec and curriculum reviews found no defects.

The full repository pytest run could not complete: the installed XGBoost
native library segfaulted in xgboost/data.py. No full-suite pass, GPU/MPS
qualification, fitting result or generalization claim is made.

* exp: preregister V5.4 single-table fit panel

* exp: record V5.4 single-table pilot launch

* exp: correct V5.4 pilot capture time

* exp: queue independent V5.4 replicates safely

* exp: record replicate queue supervisors

* exp: prepare full 120-table V5.4 fit panel

* exp: guard Mac pilot hosts from idle sleep

* fix: run V5.4 Mac pilots on MPS FP32

* exp: restart V5.4 Mac panel on MPS FP32

* exp: record MPS FP32 V5.4 pilot launch

* ops: guard MPS pilot hosts from idle sleep

* perf(v54): keep MPS readout device local

* exp: prepare V5.4 single-table panel on bd9a42d

* exp: guard V5.4 replicate queue

* exp: record fresh four-host fit launch

* exp: cap deepthought readout chunk

* exp: record deepthought chunk override

* exp: mark deepthought temporarily unavailable

* feat(v54): persist optimized readout and loss replay training

* docs(v54): preserve fitting handoff and training lessons

* exp(v54): archive replay and monitoring protocols

* docs(v54): archive owner design package in TabU-lab

---------

+++ src/tabu_lab/curriculum_v53/evaluation.py
@@ -125,7 +125,7 @@ def evaluate_probe(model, plan, probe, device, *, deadline=None):
                         # NMSE measures z-coordinate error. Raw sparse directions
                         # multiply TRAINING loss by 4, not this evaluation metric.
                         scale = score.output.facts[column.column].answers.scalar.scale
-                        errors = ((column.decoded - expected.double()) / scale).square()
+                        errors = ((column.decoded - expected.to(column.decoded.dtype)) / scale).square()
                     else:
                         errors = score.per_target[positions]
                     errors = errors.detach().cpu().tolist()
